refactor: change payment message as a shared model
